move envVars into the other env file
