+++ src/config/env.ts
@@ -1,13 +1,22 @@
-import dotenv from "dotenv";
-import path from "path";
-const nodeEnv = process.env.NODE_ENV;
-const filename = nodeEnv === "test" ? ".env.test" : ".env";
+import dotenv from 'dotenv';
+import path from 'path';
+
+export const envVars = {
+  MICROS_ENV: process.env.MICROS_ENV || 'development',
+  MICROS_SERVICE_VERSION: process.env.MICROS_SERVICE_VERSION,
+  NODE_ENV: process.env.NODE_ENV,
+  SENTRY_DSN: process.env.SENTRY_DSN,
+};
+
+const { NODE_ENV } = envVars;
+const filename = NODE_ENV === 'test' ? '.env.test' : '.env';
 const env = dotenv.config({
-  path: path.resolve(process.cwd(), filename)
+  path: path.resolve(process.cwd(), filename),
 });
 
 // TODO: add checks for environment variables here and error out if missing any
-if (env.error && nodeEnv !== "production") {
+if (env.error && NODE_ENV !== 'production') {
   throw env.error;
 }
+
 export default env.parsed || {};
